+++ addons/hr/models/hr_employee.py
@@ -10,10 +10,9 @@ from werkzeug.urls import url_encode
 from dateutil.relativedelta import relativedelta
 
 from odoo import api, fields, models, _
-from odoo.osv.query import Query
 from odoo.exceptions import ValidationError, AccessError
 from odoo.osv import expression
-from odoo.tools.misc import format_date
+from odoo.tools import format_date, Query
 
 
 class HrEmployeePrivate(models.Model):
