+++ odoo/addons/base/tests/test_osv.py
@@ -1,8 +1,8 @@
 # -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
-from odoo.osv.query import Query
 from odoo.tests.common import BaseCase
+from odoo.tools import Query
 
 
 class QueryTestCase(BaseCase):
