+++ odoo/osv/expression.py
@@ -114,23 +114,17 @@ start the server specifying the ``--unaccent`` flag.
 
 """
 import collections.abc
-import warnings
-
 import logging
 import reprlib
 import traceback
-from functools import partial
-
+import warnings
 from datetime import date, datetime, time
 
 from psycopg2.sql import Composable, SQL
 
 import odoo.modules
-from odoo.osv.query import Query, _generate_table_alias
-from odoo.tools import pycompat
-from odoo.tools.misc import get_lang
-from ..models import MAGIC_COLUMNS, BaseModel
-import odoo.tools as tools
+from ..models import BaseModel
+from odoo.tools import pycompat, Query, _generate_table_alias
 
 
 # Domain operators.
